feat: custom event 모듈 분리
